chore(theme): add token object
