+++ packages/theme/src/index.ts
@@ -1,3 +1,4 @@
 export * from './theme'
 export * from './abstract-theme'
 export * from './default'
+export * from './token-object'
