+++ packages/theme/src/token-object/index.ts
@@ -1,22 +1,8 @@
-import deepmerge from 'deepmerge'
-import { isNonEmptyArray } from './helpers/is-empty-array'
-import { nestObject } from './helpers/nest-object'
 import { createReferenceTokenResolver } from './reference-token'
-import type { Tokens, TokenObject, TokenDictionary, TokenValue } from './types'
+import { toTokenObject } from './to-token-object'
+import type { TokenObject, TokenDictionary, TokenValue } from './types'
 
-export const toTokenObject = <T extends Tokens>(tokens: T): TokenObject<T> => {
-  let result = {}
-  for (const key in tokens) {
-    const { value } = tokens[key]
-    const splitted = key.split('/')
-    if (!isNonEmptyArray(splitted)) continue
-
-    const v = nestObject(splitted, value)
-    result = deepmerge(result, v)
-  }
-
-  return result as TokenObject<T>
-}
+export { camelCaseKeys } from './helpers/changecase-keys'
 
 export const createTokenObject = <T extends TokenDictionary>(
   tokenDictionary: T,
